continue moving stuff
